# providers/edge3/src/airflow/providers/edge3/openapi/edge_worker_api_v1.yaml
...
openapi: 3.0.2
info:
  title: Airflow Edge Worker API
  version: 1.0.1
  description: |
    This is Airflow Edge Worker API - which is a the access endpoint for workers
    running on remote sites serving for Apache Airflow jobs. It also proxies internal API
    to edge endpoints.

    It is not intended to be used by any external code.

    You can find more information in AIP-69
    https://cwiki.apache.org/confluence/pages/viewpage.action?pageId=301795932
servers:
  - url: /edge_worker/v1
    description: Airflow Edge Worker API
paths:
  /worker/{worker_name}:
    patch:
      description: Set state of worker and returns the current assigned queues.
      x-openapi-router-controller: airflow.providers.edge3.worker_api.routes._v2_routes
      operationId: set_state_v2
      parameters:
      - description: Hostname or instance name of the worker
        in: path
        name: worker_name
        required: true
        schema:
          description: Hostname or instance name of the worker
          title: Worker Name
          type: string
      - description: JWT Authorization Token
        in: header
        name: authorization
        required: true
        schema:
          description: JWT Authorization Token
          title: Authorization
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkerStateBody'
              description: State of the worker with details
              examples:
              - jobs_active: 3
                queues:
                - large_node
                - wisconsin_site
                state: running
                sysinfo:
                  airflow_version: 2.10.0
                  concurrency: 4
                  edge_provider_version: 1.0.1
              title: Worker State
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerSetStateReturn'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPExceptionResponse'
          description: Bad Request
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPExceptionResponse'
          description: Forbidden
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Set State
      tags:
      - Worker
    post:
      description: Register a new worker to the backend.
      x-openapi-router-controller: airflow.providers.edge3.worker_api.routes._v2_routes
      operationId: register_v2
      parameters:
      - description: Hostname or instance name of the worker
        in: path
        name: worker_name
        required: true
        schema:
          description: Hostname or instance name of the worker
          title: Worker Name
          type: string
      - description: JWT Authorization Token
        in: header
        name: authorization
        required: true
        schema:
          description: JWT Authorization Token
          title: Authorization
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkerStateBody'
              description: State of the worker with details
              examples:
              - jobs_active: 3
                queues:
                - large_node
                - wisconsin_site
                state: running
                sysinfo:
                  airflow_version: 2.10.0
                  concurrency: 4
                  edge_provider_version: 1.0.1
              title: Worker State
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerRegistrationReturn"
                description: Registration response with the last update time of the worker.
                examples:
                - last_update: "2025-04-04T13:59:58.773870"
                title: Worker Registration
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPExceptionResponse'
          description: Bad Request
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPExceptionResponse'
          description: Forbidden
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Register
      tags:
      - Worker
...
components:
  schemas:
...
    EdgeWorkerState:
      description: Status of a Edge Worker instance.
      enum:
      - starting
      - running
      - idle
      - terminating
      - offline
      - unknown
      - maintenance mode
      - maintenance request
      - maintenance pending
      - maintenance exit
      - offline maintenance
      title: EdgeWorkerState
      type: string
    WorkerStateBody:
      description: Details of the worker state sent to the scheduler.
      type: object
      required:
      - state
      - queues
      - sysinfo
      properties:
        jobs_active:
          default: 0
          description: Number of active jobs the worker is running.
          title: Jobs Active
          type: integer
        queues:
          anyOf:
          - items:
              type: string
            type: array
          - type: object
            nullable: true
          description: List of queues the worker is pulling jobs from. If not provided,
            worker pulls from all queues.
          title: Queues
        state:
          $ref: '#/components/schemas/EdgeWorkerState'
          description: State of the worker from the view of the worker.
        sysinfo:
          description: System information of the worker.
          title: Sysinfo
          type: object
        maintenance_comments:
          description: Comments about the maintenance state of the worker.
          title: Maintenance Comments
          anyOf:
          - type: string
          - type: object
            nullable: true
      title: WorkerStateBody
...
    WorkerRegistrationReturn:
      description: The response of the worker registration.
      properties:
        last_update:
          description: Time of the last update of the worker.
          format: date-time
          title: Last Update
          type: string
      title: WorkerRegistrationReturn
    WorkerSetStateReturn:
      description: The state written in the database
      properties:
        queues:
          anyOf:
          - items:
              type: string
            type: array
          - type: object
            nullable: true
          description: List of queues the worker is pulling jobs from. If not provided,
            worker pulls from all queues.
          title: Queues
        state:
          $ref: '#/components/schemas/EdgeWorkerState'
          description: State of the worker from the view of the worker.
        maintenance_comments:
          description: Comments about the maintenance state of the worker.
          title: Maintenance Comments
          anyOf:
          - type: string
          - type: object
            nullable: true
      title: WorkerSetStateReturn
...
    HTTPExceptionResponse:
      description: HTTPException Model used for error response.
      properties:
        detail:
          anyOf:
          - type: string
          - type: object
          title: Detail
      required:
      - detail
      title: HTTPExceptionResponse
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...